fix(eth-bytecode-db): invalid chain_id type when retrieving alliance database

+++ eth-bytecode-db/eth-bytecode-db-server/tests/integration/types/verifier_alliance.rs
@@ -254,7 +254,7 @@ impl VerifierAllianceDatabaseChecker for TestCase {
         db: &DatabaseConnection,
     ) -> Option<contract_deployments::Model> {
         contract_deployments::Entity::find()
-            .filter(contract_deployments::Column::ChainId.eq(Decimal::from(self.chain_id)))
+            .filter(contract_deployments::Column::ChainId.eq(self.chain_id))
             .filter(contract_deployments::Column::Address.eq(self.address.to_vec()))
             .filter(
                 contract_deployments::Column::TransactionHash.eq(self.transaction_hash.to_vec()),
